+++ impl_core/src/main/java/io/opencensus/implcore/stats/MeasureMapInternal.java
@@ -16,7 +16,7 @@
 
 package io.opencensus.implcore.stats;
 
-import io.opencensus.stats.AttachmentValue;
+import io.opencensus.metrics.data.AttachmentValue;
 import io.opencensus.stats.Measure;
 import io.opencensus.stats.Measure.MeasureDouble;
 import io.opencensus.stats.Measure.MeasureLong;